+++ nodes/web/api-call@1.0.0/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @nanoservice-ts/api-call
 
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.26
+
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.25
+
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.24
+
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.23
+
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.22
+
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.21
+
 ## 0.1.23
 
 ### Patch Changes